# docs/swagger.yaml
definitions:
  permission.CreatePermissionRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  permission.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  permission.Permission:
    properties:
      category:
        type: string
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  permission.PermissionResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  permission.UpdatePermissionRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  role.ErrorResponse:
    properties:
      error:
        example: Описание ошибки
        type: string
    type: object
...
  role.SuccessResponse:
    properties:
      message:
        example: Операция выполнена успешно
        type: string
    type: object
  role.UpdateRequest:
    properties:
      name:
        example: Модератор
        type: string
      permission_ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      role_id:
        example: 1
        type: integer
    required:
    - name
    type: object
...
info:
  contact: {}
paths:
...
  /permissions:
    get:
      description: Получить все разрешения
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/permission.PermissionResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
      summary: Получить список разрешений
      tags:
      - permissions
    post:
      consumes:
      - application/json
      description: Создать новое разрешение
      parameters:
      - description: Данные разрешения
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/permission.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/permission.PermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
      summary: Создать разрешение
      tags:
      - permissions
  /permissions/{id}:
    delete:
      description: Удалить разрешение по ID
      parameters:
      - description: ID разрешения
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
      summary: Удалить разрешение
      tags:
      - permissions
    get:
      description: Получить разрешение по идентификатору
      parameters:
      - description: ID разрешения
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/permission.PermissionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
      summary: Получить разрешение по ID
      tags:
      - permissions
    put:
      consumes:
      - application/json
      description: Обновить данные разрешения
      parameters:
      - description: ID разрешения
        in: path
        name: id
        required: true
        type: integer
      - description: Данные разрешения
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/permission.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/permission.PermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
      summary: Обновить разрешение
      tags:
      - permissions
  /permissions/name/{name}:
    get:
      description: Получить разрешение по имени
      parameters:
      - description: Имя разрешения
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/permission.PermissionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/permission.ErrorResponse'
      summary: Получить разрешение по имени
      tags:
      - permissions
...
  /roles/{id}/permissions:
    get:
      consumes:
      - application/json
      description: Возвращает список разрешений для указанной роли
      parameters:
      - description: ID роли
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/permission.Permission'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/role.ErrorResponse'
      summary: Получить разрешения роли
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Полностью заменяет список разрешений для указанной роли
      parameters:
      - description: ID роли
        in: path
        name: id
        required: true
        type: integer
      - description: Список ID разрешений
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/role.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/role.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/role.ErrorResponse'
      summary: Обновить разрешения роли
      tags:
      - roles
...
swagger: "2.0"